+++ tests/testthat/test-plotROC.R
@@ -1,0 +1,18 @@
+p <- plotROC(SDMtune:::bm_maxent, test = SDMtune:::t)
+
+test_that("The plot contains the correct data", {
+  expect_equal(unique(p$data$set), c("Train", "Test"))
+  expect_equal(unique(p$data$pa), c(1, 0))
+  expect_named(p$data, c("set", "pa", "pred"))
+
+})
+
+test_that("The plot has the correct labels", {
+  expect_equal(p$labels$x, "False Positive Rate")
+  expect_equal(p$labels$y, "True Positive Rate")
+})
+
+test_that("error is raised", {
+  expect_error(p <- plotROC(SDMtune:::bm_maxent, val = SDMtune:::t),
+  "\"val\" argument is deprecated and will be removed in the next ")
+})